Add alerts service integration tests

+++ api/Backend/CCQuartersAPI/Responses/GetAlertsResponse.cs
@@ -13,8 +13,8 @@ namespace CCQuartersAPI.Responses
         public string UserId {  get; set; }
         public double? MaxPrice { get; set; }
         public double? MaxPricePerM2 { get; set; }
-        public double? MinArea { get; set; }
-        public double? MaxArea { get; set; }
+        public decimal? MinArea { get; set; }
+        public decimal? MaxArea { get; set; }
         public int? MinRoomCount { get; set; }
         public int? MaxRoomCount { get; set; }
         public int? Floor { get; set; }
